// front-end/src/helpers/interface.tsx
export interface JOGame {
  game_end_date: string
  game_location: string
  game_name: string
  game_season: string
  game_slug: string
  game_start_date: string
  game_year: number
  index: number
}

export interface Medals {
  discipline_title: string
  slug_name: string
  event_title: string
  event_gender: string
  medal_type: string
  participant_type: string
  participant_title: string
  country_name: string
  country_code: string
  index: number
}